вынес функции в отдельный файл
