+++ templates/main.php
@@ -1,7 +1,3 @@
-<?php
-
-?>
-
 <section class="page__main page__main--popular">
     <div class="container">
         <h1 class="page__title page__title--popular">Популярное</h1>
